--- test-suites/pool-drop-reserve.spec.ts
@@ -1,118 +1,0 @@
-import {expect} from "chai";
-import {utils} from "ethers";
-import {ProtocolErrors} from "../deploy/helpers/types";
-import {MAX_UINT_AMOUNT, ZERO_ADDRESS} from "../deploy/helpers/constants";
-// import {MockFlashLoanReceiver} from "../types";
-// import {getMockFlashLoanReceiver} from "../helpers/contracts-getters";
-import {TestEnv} from "./helpers/make-suite";
-import {loadFixture} from "@nomicfoundation/hardhat-network-helpers";
-import {testEnvFixture} from "./helpers/setup-env";
-
-describe("Pool: Drop Reserve", () => {
-  // const _mockFlashLoanReceiver = {} as MockFlashLoanReceiver;
-  let testEnv: TestEnv;
-
-  const {
-    XTOKEN_SUPPLY_NOT_ZERO,
-    VARIABLE_DEBT_SUPPLY_NOT_ZERO,
-    ASSET_NOT_LISTED,
-    ZERO_ADDRESS_NOT_VALID,
-  } = ProtocolErrors;
-
-  before(async () => {
-    testEnv = await loadFixture(testEnvFixture);
-  });
-
-  it("User 1 deposits DAI, User 2 borrow DAI variable, should fail to drop DAI reserve", async () => {
-    const {
-      deployer,
-      users: [user1],
-      pool,
-      dai,
-      weth,
-      configurator,
-    } = testEnv;
-    const depositedAmount = utils.parseEther("1000");
-    const borrowedAmount = utils.parseEther("100");
-    // setting reserve factor to 0 to ease tests, no xToken accrued in reserve
-    await configurator.setReserveFactor(dai.address, 0);
-    await dai["mint(uint256)"](depositedAmount);
-    await dai.approve(pool.address, depositedAmount);
-    await dai.connect(user1.signer)["mint(uint256)"](depositedAmount);
-    await dai.connect(user1.signer).approve(pool.address, depositedAmount);
-    await weth.connect(user1.signer)["mint(uint256)"](depositedAmount);
-    await weth.connect(user1.signer).approve(pool.address, depositedAmount);
-    await pool.supply(dai.address, depositedAmount, deployer.address, 0);
-    await expect(configurator.dropReserve(dai.address)).to.be.revertedWith(
-      XTOKEN_SUPPLY_NOT_ZERO
-    );
-    await pool
-      .connect(user1.signer)
-      .supply(weth.address, depositedAmount, user1.address, 0);
-    await pool
-      .connect(user1.signer)
-      .borrow(dai.address, borrowedAmount, 0, user1.address);
-    await expect(configurator.dropReserve(dai.address)).to.be.revertedWith(
-      VARIABLE_DEBT_SUPPLY_NOT_ZERO
-    );
-    // await pool
-    //   .connect(user1.signer)
-    //   .borrow(dai.address, borrowedAmount, 1, 0, user1.address);
-    // await expect(configurator.dropReserve(dai.address)).to.be.revertedWith(
-    //   STABLE_DEBT_NOT_ZERO
-    // );
-  });
-
-  it("User 2 repays debts, drop DAI reserve should fail", async () => {
-    const {
-      users: [user1],
-      pool,
-      dai,
-      configurator,
-    } = testEnv;
-    // expect(
-    //   await pool
-    //     .connect(user1.signer)
-    //     .repay(dai.address, MAX_UINT_AMOUNT, 1, user1.address)
-    // );
-    // await expect(configurator.dropReserve(dai.address)).to.be.revertedWith(
-    //   VARIABLE_DEBT_SUPPLY_NOT_ZERO
-    // );
-    expect(
-      await pool
-        .connect(user1.signer)
-        .repay(dai.address, MAX_UINT_AMOUNT, user1.address)
-    );
-    await expect(configurator.dropReserve(dai.address)).to.be.revertedWith(
-      XTOKEN_SUPPLY_NOT_ZERO
-    );
-  });
-
-  it("User 1 withdraw DAI, drop DAI reserve should succeed", async () => {
-    const {deployer, pool, dai, configurator, protocolDataProvider} = testEnv;
-    await pool.withdraw(dai.address, MAX_UINT_AMOUNT, deployer.address);
-    const reserveCount = (await pool.getReservesList()).length;
-    expect(await configurator.dropReserve(dai.address));
-    const tokens = await pool.getReservesList();
-    expect(tokens.length).to.be.eq(reserveCount - 1);
-    expect(tokens.includes(dai.address)).to.be.false;
-    const {isActive} = await protocolDataProvider.getReserveConfigurationData(
-      dai.address
-    );
-    expect(isActive).to.be.false;
-  });
-
-  it("Drop an asset that is not a listed reserve should fail", async () => {
-    const {users, configurator} = testEnv;
-    await expect(configurator.dropReserve(users[5].address)).to.be.revertedWith(
-      ASSET_NOT_LISTED
-    );
-  });
-
-  it("Drop an asset that is not a listed reserve should fail", async () => {
-    const {configurator} = testEnv;
-    await expect(configurator.dropReserve(ZERO_ADDRESS)).to.be.revertedWith(
-      ZERO_ADDRESS_NOT_VALID
-    );
-  });
-});
